split files by topics
