+++ src/UI.py
@@ -12,11 +12,12 @@ import time # for sleep in FAKE_AI
 import src.config as config
 import src.utils as utils
 import src.analytics as analytics
+import src.AI as AI
 
 # if active much more log output and ability to switch and select the used generation model 
 DEBUG = False
 # if active the system is not using any AI, just simulating the funcion
-FAKE_AI = False
+FAKE_AI = True
 FAKE_AI_DELAY = 5 # time how long the thread sleeps to simulate generation
 
 device = "cuda" # will be checked and set in main functions
@@ -26,28 +27,15 @@ style_details = {}
 utils.DEBUG = DEBUG
 analytics.DEBUG = DEBUG
 
-def get_all_local_models():
-    """read all local models to the system"""
-    safetensors_files = []
-    try:
-        for root, _, files in os.walk(config.get_model_folder(),followlinks=True):
-            for file in files:
-                if file.endswith('.safetensors'):
-                    relative_path = "./" + os.path.relpath(os.path.join(root, file))
-                    safetensors_files.append(relative_path)
-        print (safetensors_files)
-    except Exception as e:
-        print(e)
-    return safetensors_files
 
 def action_update_all_local_models():
     """updates the list of available models in the ui"""
-    return gr.update(choices=get_all_local_models())
+    return gr.update(choices=utils.get_all_local_models(config.get_model_folder()))
 
 def action_save_input_file(request: gr.Request, image):
     """Save the input image in a cache directory using its SHA-1 hash."""
     # deactivate the start buttons
-    if image is None or load_model()==None: return [gr.update(interactive=False), gr.update(interactive=False), ""]
+    if image is None: return [gr.update(interactive=False), gr.update(interactive=False), ""]
     
     # API Users don't have a request (by documentation)
     if not request: return
@@ -71,86 +59,13 @@ def action_save_input_file(request: gr.Request, image):
     # there is an image, activate the start buttons
     return [gr.update(interactive=True), gr.update(interactive=True), ""]
 
-def check_safety(x_image):
-    """Support function to check if the image is NSFW."""
-    # all images are SFW ;)
-    return x_image, False 
-
-IMAGE_TO_TEXT_PIPELINE = None
-def load_captioner():
-    """Load and return a image to text model."""
-    if FAKE_AI: return
-    if (IMAGE_TO_TEXT_PIPELINE!= None): return IMAGE_TO_TEXT_PIPELINE
-
-    #TODO: check stop params like 
-    captioner = pipeline("image-to-text", model="Salesforce/blip-image-captioning-base")
-    # TODO V3: change to this pipeline to query details about the image in Version 2
-    #captioner = pipeline("image-text-to-text", model="Salesforce/blip-image-captioning-base")
-    return captioner
-
 def action_describe_image(image):
     """describe an image for better inpaint results."""
     if FAKE_AI: return "ai deactivated"
     
-    captioner = load_captioner()
-
-    #TODO V3: add prompt to captioner to ask for details like how many people, background etc.
-    # messages = [
-    # {
-    #     "role": "user",
-    #     "content": [
-    #         {"type": "image", "image": "https://huggingface.co/datasets/huggingface/documentation-images/resolve/main/bee.jpg"},
-    #         {"type": "text", "text": "What do we see in this image?"},
-    #     ]
-    # }]
-    # prompt = captioner.apply_chat_template(messages)
-    #value = captioner(image, text=prompt, return_full_text=False)
-    value = captioner(image)
-
+    value = AI.describe_image(image)
     if DEBUG: print(f"Image description: {value}")
-    return value[0]['generated_text']
-
-IMAGE_TO_IMAGE_PIPELINE = None
-def load_model(model=config.get_model()):
-    """Load and return the Stable Diffusion model based on style."""
-    if FAKE_AI: return False
-
-    global IMAGE_TO_IMAGE_PIPELINE
-    if (IMAGE_TO_IMAGE_PIPELINE!= None): 
-        if DEBUG: print ("using cached model")
-        return IMAGE_TO_IMAGE_PIPELINE
-
-    try:
-        #TODO V3: support SDXL or FLux
-        if DEBUG: print(f"create pipeline for model {model}")
-
-        pipeline = None
-        if model.endswith("safetensors"):
-            if DEBUG: print("use 'from_single_file' to load model from local folder")
-            pipeline = StableDiffusionImg2ImgPipeline.from_single_file(
-                model,
-                torch_dtype=torch.float16 if device == "cuda" else torch.float32,
-                safety_checker = None, requires_safety_checker = False,
-                use_safetensors=True
-                #revision="fp16" if device == "cuda" else "",
-            )
-        else:
-            if DEBUG: print("use 'from_pretrained' option to load model from hugging face")
-            pipeline = StableDiffusionImg2ImgPipeline.from_pretrained(
-                model,
-                torch_dtype=torch.float16 if device == "cuda" else torch.float32,
-                safety_checker = None, requires_safety_checker = False)
-            
-        if DEBUG: print(f"--> pipeline initiated")
-        pipeline = pipeline.to(device)
-        if device=="cuda": pipeline.enable_xformers_memory_efficient_attention()
-        if DEBUG: print("--> pipeline created")
-        return pipeline
-    except Exception as e:
-        print(f"pipeline not be created. Error in load_model: {e}")
-        if not gr is None:
-            gr.Error(message="Error while loading the model.\nSee logfile for details.")
-        return None
+    return value
 
 
 def action_reload_model(model):
@@ -158,15 +73,10 @@ def action_reload_model(model):
     global IMAGE_TO_IMAGE_PIPELINE
     print (f"Reload model {model}")
     try:
-        # TODO (low prio): check better ways to unload a model!
-        IMAGE_TO_IMAGE_PIPELINE = None
-        if device=="cuda": torch.cuda.empty_cache()
-        # load new model
-        IMAGE_TO_IMAGE_PIPELINE = load_model(model)
-        if IMAGE_TO_IMAGE_PIPELINE is None: raise Exception("Pipeline is none")
+        AI.change_text2img_model(model=model)
         gr.Info(message=f"Model {model} loaded.", title="Model changed")
     except Exception as e:
-        print(f"model not loaded. {e}")
+        gr.Error(message=e.message)
 
 def action_generate_image(request: gr.Request, image, style, strength, steps, image_description):
     """Convert the entire input image to the selected style."""
@@ -179,14 +89,7 @@ def action_generate_image(request: gr.Request, image, style, strength, steps, im
             return None,"API forbidden"
 
         if DEBUG: print("start image generation")
-        model = load_model()
-        if image_description == None or image_description == "": image_description = action_describe_image(image)
-
-        #TODO: check image_description
-        if (not FAKE_AI and model == None): 
-            print("error: no model loaded")
-            gr.Error(message="No model loaded. Generation not available")
-            return None, ""
+        if image_description == None or image_description == "": image_description = AI.describe_image(image)
 
         sd = style_details.get(style)
         if sd == None:
@@ -207,31 +110,21 @@ def action_generate_image(request: gr.Request, image, style, strength, steps, im
         # or adapt the source image saving with thumbnail property
         image_sha1 = sha1(image.tobytes()).hexdigest()
 
-        # TODO Render previews in Version 2
-        max_size = config.get_max_size()
-        image.thumbnail((max_size,max_size))
-
-        # create a mask which covers the whole image
-        mask = Image.new("L", image.size, 255) 
         if not config.UI_show_stength_slider(): strength = sd["strength"]
         if not config.UI_show_steps_slider(): steps = sd["steps"]
 
-        if DEBUG:
-            print(f"Strength: {strength}, Steps: {steps}")
-        
         if FAKE_AI:
             result_image = utils.image_convert_to_sepia(image)
             time.sleep(FAKE_AI_DELAY)
         else:
             # Generate new picture
-            result_image = model(
+            result_image = AI.generate_image(
+                image = image,
                 prompt=prompt, 
                 negative_prompt=sd["negative_prompt"],
-                num_inference_steps=steps, 
-                image=image,
-                mask_image=mask, 
+                steps=steps, 
                 strength=strength,
-                ).images[0]
+                )
         
         # save generated file if enabled
         fn=None
@@ -384,38 +277,3 @@ def create_gradio_interface():
             )
         #end if dislaimer 
         return app
-
-# if __name__ == "__main__":
-#     app = None
-#     try:
-#         if config.read_configuration() == None: print("read configuration failed")
-#         model = config.get_model()
-#         if model.endswith("safetensors"):
-#             utils.download_file_if_not_existing(url=config.get_model_url(), local_path=model)
-
-#         print ("initializing ai models")
-
-#         device = "cuda" if torch.cuda.is_available() else "cpu"
-#         print (f"running on {device}")
-#         if FAKE_AI:
-#             print ("ai deactivated")
-#         else:
-#             IMAGE_TO_IMAGE_PIPELINE = load_model()
-#             IMAGE_TO_TEXT_PIPELINE = load_captioner()
-
-#         if config.is_analytics_enabled():
-#             analytics._DEBUG = DEBUG
-#             analytics.start()
-#         title = config.get_app_title()
-#         print (f"starting Server ""{title}""")
-#         app = create_gradio_interface()
-#         app.launch(
-#             share = config.is_gradio_shared(), 
-#             max_file_size=8*gr.FileSize.MB
-#             )
-#         analytics.stop()
-#     except Exception as e:
-#         print (e)
-#     finally:
-#         if app != None: app.close()
-#         print ("app closed")
